requirements.txt

alembic==1.13.1
flask==3.0.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
flask-sqlalchemy==3.1.1
psycopg2==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.25
Flask-Seeder==1.2.0
python-dotenv==1.0.0